search users with throttle
